# Conflicts:
#	chocoshop-portal-model/target/classes/com/chocoshop/model/Order.class
#	chocoshop-portal-model/target/maven-status/maven-compiler-plugin/compile/default-compile/createdFiles.lst
#	chocoshop-portal-model/target/maven-status/maven-compiler-plugin/compile/default-compile/inputFiles.lst
#	chocoshop-portal-web/src/main/java/com/chocoshop/controller/BasicController.java
#	chocoshop-portal-web/src/main/java/com/chocoshop/controller/OrderController.java
#	chocoshop-portal-web/src/main/resources/templates/cart.html
#	chocoshop-portal-web/src/main/resources/templates/navbar.html
#	chocoshop-portal-web/target/maven-status/maven-compiler-plugin/compile/default-compile/inputFiles.lst

+++ chocoshop-portal-model/src/main/java/com/chocoshop/model/Order.java
@@ -17,7 +17,7 @@ public class Order implements Serializable {
     @Id
     private String orderUuid;
     @DecimalMin("0")
-    private Double orderPayment;
+    private BigDecimal orderPayment;
     @Range(min = 0, max = 2)
     private BigDecimal orderPaymentType;
     @Range(min = 0, max = 7)
@@ -40,11 +40,11 @@ public class Order implements Serializable {
         this.orderUuid = orderUuid;
     }
 
-    public Double getOrderPayment() {
+    public BigDecimal getOrderPayment() {
         return orderPayment;
     }
 
-    public void setOrderPayment(Double orderPayment) {
+    public void setOrderPayment(BigDecimal orderPayment) {
         this.orderPayment = orderPayment;
     }
